Several fixes while testing with Ansible 2.14.6 (#474)
* Remove unsupported warn parameter from tasks

The warn parameter isn't supported anymore with newer Ansible versions.

* Bugfix grub config file content

The lookup module is executed on the controller and not on the remote host. Replace the lookup with a slurp.

* Bugfix BOM aggregate only find local bom folders

* Refactor curl command for Azure instance metadata to Ansible uri module

The Ansible uri module recognizes the json output returned by the azure metadata instance api. No converting to json is therefore needed anymore.

* Refactor anydb_node dns_info_vms output

* Remove how file